% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{economy_patterns}
\alias{economy_patterns}
\title{Economy Patterns}
\format{
A data frame with the following columns:
\describe{
\item{pattern}{Regular expression pattern for matching economy names}
\item{iso3}{ISO 3166-1 alpha-3 code}
\item{iso2}{ISO 3166-1 alpha-2 code}
}
}
\source{
Data manually prepared by Teal L. Emery
}
\usage{
economy_patterns
}
\description{
A dataset containing patterns for matching economy names.
}
\keyword{datasets}
