% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_economy_patterns.R
\name{list_economy_patterns}
\alias{list_economy_patterns}
\title{Economy Patterns Data}
\usage{
list_economy_patterns()
}
\value{
A character vector containing regular expression patterns for economic indicators
}
\description{
This dataset contains regular expression patterns for identifying economic indicators.
It is accessible through this function.
}
\examples{
patterns <- list_economy_patterns()

}
\keyword{internal}
