% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_economies.R
\name{match_economy_ids}
\alias{match_economy_ids}
\title{Match Economy Ids}
\usage{
match_economy_ids(names, codes = NULL, warn_ambiguous = TRUE)
}
\arguments{
\item{names}{Character vector of economy names to standardize}

\item{codes}{Optional character vector of economy codes}

\item{warn_ambiguous}{Logical; whether to warn about ambiguous matches}
}
\value{
A vector of economy ids
}
\description{
Given vectors of names and codes, match them against a list of patterns and
return a vector of economy ids.
}
\keyword{internal}
