% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_entity.R
\name{match_entity_ids}
\alias{match_entity_ids}
\title{Match entity Ids}
\usage{
match_entity_ids(names, codes = NULL, warn_ambiguous = TRUE)
}
\arguments{
\item{names}{Character vector of entity names to standardize}

\item{codes}{Optional character vector of entity codes}

\item{warn_ambiguous}{Logical; whether to warn about ambiguous matches}
}
\value{
A vector of entity ids
}
\description{
Given vectors of names and codes, match them against a list of patterns and
return a vector of entity ids.
}
\keyword{internal}
