% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_entities.R
\name{standardize_entities}
\alias{standardize_entities}
\title{Standardize entity Names and Codes}
\usage{
standardize_entities(
  data,
  name_col,
  code_col = NULL,
  output_cols = c("entity_id", "entity_name", "entity_type"),
  default_entity_type = NA_character_,
  warn_ambiguous = TRUE,
  overwrite = TRUE,
  warn_overwrite = TRUE
)
}
\arguments{
\item{data}{A data frame or tibble containing entity names to standardize}

\item{name_col}{Name of the column containing entity names}

\item{code_col}{Optional name of the column containing entity codes}

\item{output_cols}{Character vector specifying desired output columns.
Options are "entity_name", "entity_id", "entity_type", "iso3c", "iso2c"}

\item{default_entity_type}{Character; the default entity type to use if not
specified in the data. Options are "country", "institution", or
"aggregate". Defaults to NA. Will be ignored if output_cols do not
include "entity_type".}

\item{warn_ambiguous}{Logical; whether to warn about ambiguous matches}

\item{overwrite}{Logical; whether to overwrite existing entity_* columns}

\item{warn_overwrite}{Logical; whether to warn when overwriting existing
entity_* columns. Defaults to TRUE.}
}
\value{
A data frame with standardized entity information merged with the
input data
}
\description{
Standardizes entity names and codes in a dataset by matching them against a
predefined list of patterns and ISO codes. Handles aggregate entities.
}
\examples{
\dontrun{
df <- data.frame(entity = c("United States", "China"))
standardize_entity(df, name_col = entity)
}

}
