% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_economies.R
\name{standardize_with_custom_names}
\alias{standardize_with_custom_names}
\title{Apply Custom Name Mappings}
\usage{
standardize_with_custom_names(x, custom_names)
}
\arguments{
\item{x}{Character vector of economy names}

\item{custom_names}{Named vector of custom name mappings}
}
\value{
Character vector with applied custom name mappings
}
\description{
Applies custom name mappings to a vector of economy names.
}
\keyword{internal}
