% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_cases}
\alias{test_cases}
\title{Economy Patterns Test Cases}
\format{
A data frame with the following columns:
\describe{
\item{economy_name}{Economy name}
\item{variant_names}{List of variant names}
}
}
\source{
DeepSeek-V3 API
}
\usage{
test_cases
}
\description{
A dataset containing test cases for the economy patterns.
}
\keyword{internal}
