% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_economies.R
\name{try_iso_match}
\alias{try_iso_match}
\title{Try ISO Code Match}
\usage{
try_iso_match(name, patterns)
}
\arguments{
\item{name}{Character string of economy name}

\item{patterns}{Tibble of matching patterns}
}
\value{
List containing matched economy information or NULL
}
\description{
Attempts to match an economy name directly against ISO codes.
}
\keyword{internal}
