% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_economies.R
\name{try_regex_match}
\alias{try_regex_match}
\title{Try Regex Pattern Match}
\usage{
try_regex_match(name)
}
\arguments{
\item{name}{Character string of economy name or code}
}
\value{
Character vector of economy_ids
}
\description{
Attempts to match a string and return matching economy_id(s) using regex
patterns from economy_patterns.
}
\keyword{internal}
